backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
mediapipe==0.10.32
opencv-python-headless==4.11.0.86
numpy==2.2.3
Pillow==11.1.0
